requirements.txt
asgiref==3.5.0
attrs==21.4.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
